[TASK] Refactor Unit Test
As the ResourceServer needs absolute file path to key files,
Configuration class returns now the system internal correct file path.
Unit Test expected the values from LocalConfiguration "as is", which now
is no longer given. In this case, we only check, if the values are given
to configuration and handled correct