[TASK] Add `docker-compose` workaround to GitHub Action
`Build/Scripts/runTests.sh` needs to be further streamlined
to the more modern `docker compose` less setup, which marks
an additional workload out of the scope of this change. But
GitHub removed the old `docker-compose` script in favour of
the newer `docker compose` plugin variant and breaks current
`Build/Scripts/runTests.sh` execution.

To mitigate the issue for now the GitHub action workflows
are now creating a bsah script to dispatch from the script
to the plugin.