[BUGFIX] Add missing allowed class for OAuth request unserilization
Restricting allowed classes when unserializing strings
into a classe structure is a security restricting to
mitigate exploiting abitary classes. Not listed classes
will be deserialized as an incomplete PHP class, ending
into an TypeError when passed to a method with strict
native type declaration.

This change adds the required custom scrope class to
the allowed list when trying to unserialize the oauth
request. That mitigates the TypeError now when doing
a successfull login.